+++ ui/pages/tasks.vue
@@ -330,7 +330,7 @@ import { useStorage } from '@vueuse/core'
 import { CronExpressionParser } from 'cron-parser'
 import { request, sleep } from '~/utils/index'
 import { encode } from '~/utils/importer'
-import type { task_item, exported_task, error_response } from '~/@types/tasks'
+import type { task_item, exported_task, error_response } from '~/types/tasks'
 
 const box = useConfirm()
 const toast = useNotification()
