+++ ui/types/conditions.d.ts
