+++ ui/types/config.d.ts
